+++ xrpl/models/amounts/__init__.py
@@ -11,6 +11,7 @@ from xrpl.models.amounts.amount import (
     is_mpt,
     is_xrp,
 )
+from xrpl.models.amounts.clawback_amount import ClawbackAmount
 from xrpl.models.amounts.issued_currency_amount import IssuedCurrencyAmount
 from xrpl.models.amounts.mpt_amount import MPTAmount
 
@@ -22,4 +23,5 @@ __all__ = [
     "is_mpt",
     "get_amount_value",
     "MPTAmount",
+    "ClawbackAmount",
 ]